requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
altgraph==0.17.4
anyio==4.4.0
attrs==23.2.0
beautifulsoup4==4.12.3
Brotli==1.1.0
bs4==0.0.2
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
comtypes==1.4.5
contourpy==1.2.1
cycler==0.12.1
fonttools==4.53.0
frozenlist==1.4.1
g4f==0.3.2.2
gevent==24.2.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.25.2
idna==3.7
Jinja2==3.1.4
keyboard==0.13.5
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.9.0
MouseInfo==0.1.3
multidict==6.0.5
Nuitka==2.3.1
numpy==2.0.0
opencv-python==4.10.0.84
ordered-set==4.1.0
outcome==1.3.0.post0
packaging==24.1
pefile==2023.2.7
pillow==10.4.0
PyAutoGUI==0.9.54
pycparser==2.22
pycryptodome==3.20.0
pydantic==1.10.17
pygame==2.6.0
PyGetWindow==0.0.9
pyinstaller==6.8.0
pyinstaller-hooks-contrib==2024.7
PyMsgBox==1.0.9
pyparsing==3.1.2
pyperclip==1.8.2
pypiwin32==223
PyRect==0.2.0
PyScreeze==0.1.30
PySocks==1.7.1
pystray==0.19.5
python-dateutil==2.9.0.post0
pyttsx3==2.90
pytweening==1.2.0
PyWavefront==1.3.3
pywin32==306
pywin32-ctypes==0.2.2
requests==2.32.3
selenium==4.21.0
setuptools==70.1.1
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
stalcraft-api==1.1.1
trio==0.25.1
trio-websocket==0.11.1
truststore==0.9.1
typing_extensions==4.12.2
urllib3==2.2.2
websocket==0.2.1
websocket-client==1.8.0
wsproto==1.2.0
yarl==1.9.4
zope.event==5.0
zope.interface==7.0.1
zstandard==0.22.0
